Try to repair unbroken link contents.
Customize the contents of the repaired link to fit a more appropriate
content, replacing the contents of 'index.html' (that are already OK
in 'index.html'), limiting to a short notice of the new location of
the resource. Assuming the newly pointed resource is a decent
replacement of the now missing blog post.